리눅스 빌드 해보고 일단 보관

+++ src/qg/qgwin_mesg.c
@@ -11,7 +11,7 @@
 #ifdef _DEBUG
 static struct WindowsMessageMap
 {
-	UINT				mesg;
+	size_t				mesg;
 	const char*			str;
 }
 s_map[] =
